refactor virtualized list to own file
